Fault rock production section
Added the description of the logic behind the model predicting the amount of fault rock produced in a fault zone. I have run into a bit of an issue regarding how to evaluate the difference between the volume under the initial and final surface. The Change is not one-to-one. Moreover, it fundamentally depends on how much the mean plane has gone down.